Refactor TodayWord Component and Implement Custom Hook
- Rozdzielono logikę komponentu TodayWord na niestandardowy hak o nazwie useTodayWord.
- Zaktualizowano komponent TodayWord, aby wykorzystać niestandardowy hak do zarządzania stanem i działaniami.
- Zapewniono, że niestandardowy hak obsługuje pobieranie danych, aktualizacje stanu i interakcje, takie jak pokazywanie słowa, mówienie słowa i oznaczanie słowa jako nauczonego.
- Add typs in baseWord, transWord, idWord.
